move clientApi to src dir
