+++ src/clientApi.ts
@@ -1,17 +1,17 @@
-import { type AddBooksToCategoryParams } from "./api/category/addBooks";
+import { type AddBooksToCategoryParams } from "./pages/api/category/addBooks";
 
 import { post, get } from "y/utils/request";
-import { type CreateCategoryParam } from "./api/category/create";
-import { type RemoveCategoryParam } from "./api/category/remove";
-import { type LoginParam } from "./api/user/login";
-import { type DeleteBookParam } from "./api/book/destroy";
-import { type CreateMarkParams } from "./api/mark/create";
-import { type GetMarkQuery } from "./api/mark";
-import { type UpdateMarkParams } from "./api/mark/update";
+import { type CreateCategoryParam } from "./pages/api/category/create";
+import { type RemoveCategoryParam } from "./pages/api/category/remove";
+import { type LoginParam } from "./pages/api/user/login";
+import { type DeleteBookParam } from "./pages/api/book/destroy";
+import { type CreateMarkParams } from "./pages/api/mark/create";
+import { type GetMarkQuery } from "./pages/api/mark";
+import { type UpdateMarkParams } from "./pages/api/mark/update";
 import type * as Prisma from "@prisma/client";
-import { type CreateUserParams } from "./api/user/create";
-import { type DeleteUserParams } from "./api/user/destroy";
-import { type PasswordParams } from "./api/user/password";
+import { type CreateUserParams } from "./pages/api/user/create";
+import { type DeleteUserParams } from "./pages/api/user/destroy";
+import { type PasswordParams } from "./pages/api/user/password";
 
 export function apiAddBooksToCategory(data: AddBooksToCategoryParams[]) {
   return post("/api/category/addBooks", data);
