+++ src/components/pages/reader/epubReader.tsx
@@ -7,7 +7,7 @@ import {
   HighlightEditor,
   getColorsValue,
 } from "y/components/highlightEditor";
-import { addMark, removeMark, apiUpdateMark } from "../../../pages/clientApi";
+import { addMark, removeMark, apiUpdateMark } from "../../../clientApi";
 import { getElementHeading } from "../../../pages/reader/index";
 import type * as Prisma from "@prisma/client";
 import { MarkType } from "y/utils/constants";
